# docs/swagger.yaml
basePath: /
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  schema.GetSiteLegalInfoResp:
    properties:
      privacy_policy_original_text:
        type: string
      privacy_policy_parsed_text:
        type: string
      terms_of_service_original_text:
        type: string
      terms_of_service_parsed_text:
        type: string
    type: object
...
info:
  contact: {}
  description: Apache Answer API
  title: Apache Answer
paths:
...
  /answer/api/v1/siteinfo/legal:
    get:
      description: get site legal info
      parameters:
      - description: legal information type
        enum:
        - tos
        - privacy
        in: query
        name: info_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.GetSiteLegalInfoResp'
              type: object
      summary: get site legal info
      tags:
      - site
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"